Made Simulator an abstract class
Simulator is now an abstract class, as there are many things that
we may want to simulate. Might not need to be an abstract class, but
it is nice to have some common code across types of simulations that
I wont need to redo. Also maybe at some point I will want to run
easily run an arbitrary amount of Simulations at a time and would like
to loop throw a list to update them.